Fix Linux case-sensitivity crash and harden manage pages

+++ app/Http/Controllers/Admin/BrandController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers\Admin;
 use App\Models\brand;
 use App\Http\Controllers\Controller;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\DB;
 
 class BrandController extends Controller
 {
@@ -31,7 +32,7 @@ class BrandController extends Controller
 
     public function edit($id)
     {
-        $brand = Brand::where('id', $id)->first();
+        $brand = Brand::findOrFail($id);
         return view('admin.brand.edit',["brand"=> $brand]);
     }
 
@@ -41,7 +42,7 @@ class BrandController extends Controller
             "name" => "required",
         ]);
 
-        $brand = Brand::where('id', $id)->first();
+        $brand = Brand::findOrFail($id);
         $brand->brand_name = $request->name;
         $brand->save();
         return redirect('admin/brand')->with('msg','Brand Updated!!!');
@@ -49,14 +50,17 @@ class BrandController extends Controller
 
     public function delete($id)
     {
-        $brand = Brand::where('id', $id)->first();
+        $brand = Brand::findOrFail($id);
+        if (DB::table('product')->where('brand_id', $id)->exists()) {
+            return redirect('admin/brand')->with('msg', 'Cannot delete: brand is in use by products.');
+        }
         $brand->delete();
         return redirect('admin/brand')->with('msg','Brand Deleted');
     }
 
     public function toggle($id)
     {
-        $brand = Brand::where('id', $id)->first();
+        $brand = Brand::findOrFail($id);
         if($brand->status == 1){
             $brand->status = 0;
             $brand->save();
